refactor codebase to use controllers
