+++ server/controllers/recipes/index.js
@@ -1,0 +1,13 @@
+import createRecipe from './addRecipe';
+import getAllRecipe from './getAllRecipe';
+import updateRecipe from './updateRecipe';
+import deleteRecipe from './deleteRecipe';
+
+const recipesController = {
+  getAllRecipe,
+  createRecipe,
+  deleteRecipe
+};
+
+
+export default recipesController;
