+++ server/controllers/recipes/updateRecipe.js
@@ -1,0 +1,7 @@
+import { Recipe } from '../../models/index';
+
+const updateRecipe = (req, res, next) => {
+
+};
+
+export default updateRecipe;
