+++ server/controllers/users/index.js
@@ -1,0 +1,10 @@
+import login from './login';
+import signup from './signup';
+
+const usersController = {
+  login,
+  signup
+};
+
+
+export default usersController;
